requirements.txt
asgiref==3.7.2
astroid==3.0.2
black==23.12.1
blinker==1.7.0
boto3==1.34.29
botocore==1.34.29
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.7.1
dill==0.3.7
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-allow-cidr==0.7.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environ==1.0
Flask==3.0.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pylint==3.0.3
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.1
whitenoise==6.6.0
zipp==3.17.0
